Utökat med fler tester och funktionalitet för Quest, Player, GameCharacter. Ändrat från int till double för Player/gamecharacter Skapat nya getmetoder

+++ RogueLikeMavenBuild/src/main/java/rougelike/combat/Action.java
@@ -7,7 +7,7 @@ public class Action {
     public static void hit(GameCharacter actionChar, GameCharacter reactiveChar){
 
         int baseDamage = 10;
-        int totalDamage = baseDamage * actionChar.getStrength();
+        double totalDamage = baseDamage * actionChar.getStrength();
 
         double effect = reactiveChar.getHealth() - totalDamage;
         reactiveChar.setHealth(effect);
